Changed type of elements from int to long long. Created another matlab script for generating matrices

+++ DiscreteCosineTransform/Transform.h
@@ -2,7 +2,7 @@
 
 #include <vector>
 
-typedef std::vector<std::vector<int>> Matrix;
+typedef std::vector<std::vector<long long>> Matrix;
 
 /*
 	Class for storing and calculating phases of DCT.
